+++ packages/components/tests/ScreenReaderShortcutTest.js
@@ -1,10 +1,10 @@
-jest.unmock( "../ScreenReaderShortcut" );
+jest.unmock( "../a11y/ScreenReaderShortcut" );
 jest.unmock( "prop-types" );
 
 import React from "react";
 import ReactShallowRenderer from "react-test-renderer/shallow";
-import ScreenReaderShortcut from "../ScreenReaderShortcut";
-import Styles from "../Styles";
+import ScreenReaderShortcut from "../a11y/ScreenReaderShortcut";
+import Styles from "../a11y/Styles";
 import { shallow } from "enzyme";
 
 describe( "ScreenReaderShortcut", () => {
