+++ packages/components/tests/ScreenReaderTextTest.js
@@ -1,10 +1,10 @@
-jest.unmock( "../ScreenReaderText" );
+jest.unmock( "../a11y/ScreenReaderText" );
 jest.unmock( "prop-types" );
 
 import React from "react";
 import ReactShallowRenderer from "react-test-renderer/shallow";
-import ScreenReaderText from "../ScreenReaderText";
-import Styles from "../Styles";
+import ScreenReaderText from "../a11y/ScreenReaderText";
+import Styles from "../a11y/Styles";
 
 describe( "ScreenReaderText", () => {
 	const renderer = new ReactShallowRenderer();
